# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  durationMinutes: number;
  price: number;
  category: string;
  featured: boolean;
}

export interface StylistScheduleEntry {
  /**
     * @minimum 0
     * @maximum 6
     */
  dayOfWeek: number;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  openTime: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  closeTime: string;
}

export interface Stylist {
  id: number;
  name: string;
  role: string;
  bio: string;
  initials: string;
  accent: string;
  schedule: StylistScheduleEntry[];
}

export interface StylistScheduleUpdate {
  /** @maxItems 21 */
  schedule: StylistScheduleEntry[];
}

export interface Availability {
  stylistId: number;
  date: string;
  slots: string[];
}

export interface Appointment {
  id: number;
  serviceId: number;
  serviceName: string;
  stylistId: number;
  stylistName: string;
  customerName: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  /** @nullable */
  notes: string | null;
  status: string;
  createdAt: string;
}

export interface AppointmentInput {
  serviceId: number;
  stylistId: number;
  /** @minLength 2 */
  customerName: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  date: string;
  time: string;
  /** @nullable */
  notes?: string | null;
}

export interface SalonSummary {
  rating: number;
  reviewCount: number;
  nextAvailable: string;
  neighborhood: string;
  hours: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetAvailabilityParams = {
date: string;
stylistId: number;
serviceId: number;
};

export type ListAppointmentsParams = {
email: string;
};