# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: salon
    description: Salon booking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /services:
    get:
      operationId: listServices
      tags: [salon]
      summary: List salon services
      responses:
        "200":
          description: Available salon services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /stylists:
    get:
      operationId: listStylists
      tags: [salon]
      summary: List salon stylists
      responses:
        "200":
          description: Available stylists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stylist"
  /stylists/{stylistId}/schedule:
    patch:
      operationId: updateStylistSchedule
      tags: [salon]
      summary: Update an employee's working schedule
      security:
        - SalonManager: []
      parameters:
        - name: stylistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StylistScheduleUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stylist"
        "400":
          description: Invalid or overlapping schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Manager access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /availability:
    get:
      operationId: getAvailability
      tags: [salon]
      summary: Get appointment time slots for an employee, service, and date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: stylistId
          in: query
          required: true
          schema:
            type: integer
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots for the selected employee
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Availability"
        "400":
          description: Invalid date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments:
    get:
      operationId: listAppointments
      tags: [salon]
      summary: List appointments for a guest email
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Guest appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [salon]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid or unavailable appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /salon-summary:
    get:
      operationId: getSalonSummary
      tags: [salon]
      summary: Get a lightweight salon summary
      responses:
        "200":
          description: Salon summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalonSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
        category:
          type: string
        featured:
          type: boolean
      required:
        - id
        - name
        - description
        - durationMinutes
        - price
        - category
        - featured
    Stylist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        bio:
          type: string
        initials:
          type: string
        accent:
          type: string
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/StylistScheduleEntry"
      required:
        - id
        - name
        - role
        - bio
        - initials
        - accent
        - schedule
    StylistScheduleEntry:
      type: object
      properties:
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        openTime:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        closeTime:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
      required:
        - dayOfWeek
        - openTime
        - closeTime
    StylistScheduleUpdate:
      type: object
      properties:
        schedule:
          type: array
          maxItems: 21
          items:
            $ref: "#/components/schemas/StylistScheduleEntry"
      required:
        - schedule
    Availability:
      type: object
      properties:
        stylistId:
          type: integer
        date:
          type: string
          format: date
        slots:
          type: array
          items:
            type: string
      required:
        - stylistId
        - date
        - slots
    Appointment:
      type: object
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        stylistId:
          type: integer
        stylistName:
          type: string
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serviceId
        - serviceName
        - stylistId
        - stylistName
        - customerName
        - email
        - phone
        - date
        - time
        - notes
        - status
        - createdAt
    AppointmentInput:
      type: object
      properties:
        serviceId:
          type: integer
        stylistId:
          type: integer
        customerName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - serviceId
        - stylistId
        - customerName
        - email
        - phone
        - date
        - time
    SalonSummary:
      type: object
      properties:
        rating:
          type: number
        reviewCount:
          type: integer
        nextAvailable:
          type: string
        neighborhood:
          type: string
        hours:
          type: string
      required:
        - rating
        - reviewCount
        - nextAvailable
        - neighborhood
        - hours
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
  securitySchemes:
    SalonManager:
      type: apiKey
      in: header
      name: X-Salon-Manager