# lib/api-zod/src/generated/types/stylistScheduleEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StylistScheduleEntry {
  /**
     * @minimum 0
     * @maximum 6
     */
  dayOfWeek: number;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  openTime: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  closeTime: string;
}