# lib/api-zod/src/generated/types/stylistScheduleUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StylistScheduleEntry } from './stylistScheduleEntry';

export interface StylistScheduleUpdate {
  /** @maxItems 21 */
  schedule: StylistScheduleEntry[];
}